Ignore multiple ActiveRecord::ConcurrentMigrationError during deployments
We have multiple pods for the datastore running in each of our
namespaces. Each time we deploy we throw an ActiveRecord::ConcurrentMigrationError
as the first pod available will be the one that manages to successfully
run the migration. Any following pods will just throw an error.

We can mitigate this by catching the ActiveRecord::ConcurrentMigrationError
and this should not trigger any Sentry alerts.