Disable wrap parameters and a bit of refactor
And a bit of refactor to get rid of some unpermitted params errors.

Removed a `bad_request` condition because this cannot trigger, as the only validation performed on the model is presence of `service_slug` attribute and this will always be present as it is part of the route, and without it the create action can't be hit.